fix: load more kernel modules in sloppy hostonly mode
When hostonly_mode is set to "sloppy" instmods consults the host for currently loaded kernel modules and
refuses to copy the Linux kernel module into the generated initramfs unless the kernel
module is already loaded on the host.

Let's set hostonly explicitly to a better default before calling instmods.

For compatibility, we should not change the default in the instmods function itself.