fix(modules): free up range 00-09 to out of tree dracut modules
This PR does not change the ascending sort order of the modules.

It adds 10 to all dracut modules in the range of 00-10 to move it up
to the 10-20 range.

The motivation for this PR is to have namespace (dracut module numbers)
reserved at the beginning (similar to the end) of the range.

Fixes https://github.com/dracut-ng/dracut-ng/issues/538.