fix(systemd-repart): allow partition format
systemd-repart is capable not only of creating a partition, but also of
formatting it. According with repart.d, it is able to create the
following filesystems: ext4, btrfs, xfs, vfat, erofs and squashfs.

Add support in the systemd-repart module for the underlying tools to
allow systemd-repart to format the partition.

Failure to do so would make systemd-repart initramfs unit fail, if
Format= option is provided in a repart.d config file.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>