test: add --pbkdf pbkdf2 to luksFormat commands
Make the test work with recent release (v2.8.0) of cryptsetup.

This PR keeps memory requirements for LUKS unlocking minimal,
allowing testcases run in system with minimal memory.

The storage stack and boot process that is being tested,
remains exactly the same.

Default for LUKS2 is now Argon2 (memory-hard KDF,
so it means that it requires some amount of memory).

Fixes: https://github.com/dracut-ng/dracut-ng/issues/1389