ci: disable running arm64 tests on Fedora
Issue filed https://github.com/dracut-ng/dracut-ng/issues/1406

Why ? The primary purpose of the CI is to help contributors to highlight
regressions caused by PRs. Leaving tests broken caused by
changes in distributions and not by dracut changes would make it
harder to contribute to dracut.

Of course ideally we want to respond changes in distributions,
but for this we realy on Linux distribution contributors.