feat: split compressed files as an option
A uncompressed cpio is added before the main cpio but after the early
cpio to hold the compressed files. This provides a time save without
much bloat in the produced file.

lsinitrd.sh is updated to handle the new cpio file. It is now able to
process an arbitrary amount of preceeding cpio files if the first one
is detected as an early_cpio or a nocompress_cpio. Detection code still
needs work to handle the new Debian format, which is like ours but
without any file flagging the segments as such.

It is theoretically possible to put the uncompressed cpio after the main
file, but (1) padding will be required (2) lsinitrd.sh will need a lot
more work.