fix(dracut.sh,lsinitrd.sh): use obj{copy,dump} set by toolchain
If dracut is invoked in a context where the toolchain is configured
to use a certain objcopy implementation, use that one. This may be
the case if dracut is invoked via systemd's kernel-install in a
Gentoo ebuild on a llvm-only system.

Also ensure that OBJCOPY and OBJDUMP contains the full path since we
manipulate PATH later on in the script.

Closes: https://github.com/dracut-ng/dracut-ng/issues/1424
Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>