fix(generic.conf): increase ordering for generic.conf
In the unlikely case that both hostonly and generic config
is installed generic should take an explicit precedence.

Currently both hostonly and generic config has the same ordering
number (10), so the precedence is based on the rest of the name
and not based on the ordering number.

Now that hostonly is the default configuration that gets installed
by default, it is particularly important for generic has higher
precedence as generic is likely only installed as an opt-in
configuration and not as the "default".