fix(dracut): consolidate reporting running in a container
dinfo call to report running in a container does not work as expected
as dinfo is not yet defined in the first code block.

In addition we're already reporting running in a container in some
other conditions, which would lead to reporting it twice.

Consolidate reporting running in a container into one single place
and make it work, and reporting is as information and not as a warning
as using dracut inside container is common.

Reordered two code blocks to enable the usage of dinfo to simplify
reporting.

Follow-up to 2b2debd .