fix(dracut): only call uname -r if it is safe to do
Change the priority order to determine the kernel version of it is
not specified on the command line.

Only call uname -r if it is safe to do, meaning if it is confirmed
that dracut is not called from within a container.

This commit is required for the tests to continue to pass after the
other changes in this PR.

Follow-up to 2b2debd.