fix(dracut): default to initrd.img-${kernel} on Debian/Ubuntu
Debian/Ubuntu carries a patch to change the default `initrdname` to
`initrd.img-${kernel}`. This is done to keep the name the same as
initramfs-tools which has been the default for a long time. This change
will be carried for quite some time to ease the switch from
initramfs-tools to dracut.

So better include this default change for Debian/Ubuntu upstream.