fix: place dracut generated cmdline conf files in the 10-49 range
The main motivation of this PR is to make it easier to reason
about the ordering numbers of dracut generated cmdline conf files.

Prefix almost all dracut generated cmdline conf file in the 20 ordering
as the ordering is usually not significant for these cmdline conf file.

Document that 50-59 is the recommended ordering range for
3rd party cmdline conf files to make it consistent with e.g.
3rd party dracut modules recommendation.

Fixes https://github.com/dracut-ng/dracut-ng/issues/1489