feat(configure): autoconfigure configprofile
The original motivation of this PR was to set the openSUSE config
inside the openSUSE CI container.

This commit will autoconfigure those configprofiles that are
maintained upstream for fedora, debian and opensuse based distro's.

One notable change from this commit is that it would now allow testing
hostonly_cmdline=yes configuration which is currently untested
on the CI.