ci(integration-extra): disable UEFI test on debian:sid
This test has been failing on debian:sid for a while and passing on
all other test containers.

I was unable to find a root cause and what is even more interesting
is that the test is passing on debian:latest (trixie).

At this point of time debian:sid and debian:latest (trixie) should
be very similar, so there is a decent chance that this is a
distribution issue and not a dracut issue.

This is the last remaining failure to make the integration-extra
GA green.