fix(dracut): use the same systemd config as on the host
Dracut is very opinionated when it comes to configuring systemd
in /etc/conf.d/system.conf.

Currently dracut allows setting systemdsystemunitdir differently
from the host which likely a misconfiguration that leads to errors.

Remove systemdsystemunitdir and a few other variables from
/etc/conf.d/system.conf and only keep systemdutildir,
which is used to configure the patch for udevd.