feat(kernel-modules-apply): make kernel modules from initramfs available
Although the kernel module for the filesystem is in the initramfs,
it might not loaded in the initramfs, because it is not needed to mount
the root filesystem.

This feature arranges for the kernel modules in the initramfs to be
placed into /lib/modules.

When booting with an external-to-rootfs kernel you need to ensure that
/lib/modules contains any necessary modules not already loaded.