fix(base): set udev log level properly
since udev logging functionality was added in af67d62 to dracut,
in commit 25e773eeb4 in systemd, udev has since migrated from its own
logging system to the logging system found in the rest of systemd, which
uses a different environment variable for setting the log level.  this
will allow for systems to truely be quiet during boot (i.e., before
showing a splash or before the init system begins being loud)

instead of using the systemd-specific environment variable, we set it in
udev.conf so that the change also applies to eudev-and-friends

hilariously, udev did the migration the day after dracut added support!
if systems were kept up to date with live branches, that means the
original change would have only worked for one day.

Signed-off-by: Mason Rocha <mrocha@turretllc.us>