fix: make logging functions return 1 only on invalid log level
Change the internal logging functions to return 0 if given a valid log
level and 1 if given an invalid log level, and document this.

The previously documented return value was inaccurate: `_do_dlog` would
not always return 0 as claimed, but rather would return 1 if
`lvl <= kmsgloglvl` due to using `&&` rather than an `if` block for the
final conditional.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>