feat(systemd-coredump): save coredumps to journal
This is the current process to get the core dump file of a process that crashed
in the initrd:

- Set `rd.break` in the kernel cmdline to enter the emergency shell.
- The core dump file should be located in /var/lib/systemd/coredump/, but /var
is not persistent in the initrd, and /sysroot is mounted read-only. So, the user
has to manually `mount -o remount,rw /sysroot` and then
`cp /var/lib/systemd/coredump/* /sysroot/var/lib/systemd/coredump`.

To make this process much easier, store in the journal coredumps generated in
the initrd.