test: print end of client tests even in failure case
If client tests started and failed with a command failing, the test will
exit and the ERR trap will execute `print_test_result`. But the client
test end will not be printed.

So call `client_test_end` in `print_test_result` to print a client test
end result in case a client test was in progress.