ci: explicitly skip tests are not meant to be supported
The motivation of this commit is to not run tests on the CI
that would be skipped. This saves some CI resources and
enables to get back to restore green CI, now that
skipped tests are reported as failures after f120814.

This PR does not change code coverage. This is simply
a different and more efficient way to skip tests on the CI.

Fixes: https://github.com/dracut-ng/dracut-ng/issues/1886