fix(dracut-functions): make sure UUID read from udev is correct
Sometimes the information read from udev may be outdated, for example a
disk has just got repartitioned. We don't know why this happens, but it
is better to make sure the unique identifiers are correct before returning
from get_persistent_dev(), preventing it from waiting an non-existant
device later on.

We compare the ID provided by get_persistent_dev() and the probed
result from blkid. If they don't match, we use the latter one.

This inconsistency is observed during initramfs generation stage of the
OS installation. It results in an unbootable OS, stuck at initrd trying
to wait a nonexistant UUID.

Signed-off-by: Xinhui Yang <cyan@cyano.uk>