feat(dracut-systemd): preserve the content of /var/lib/systemd
Persist the content of /var/lib/systemd in memory during initramfs
execution by linking it to /run/initramfs/systemd .

One of the motivation of this PR is to preserved coredumps
when systemd-coredump dracut module is included in the initrfamfs.

In general, this approach would requires less customization of
systemd configuration for dracut.