fix(nfs): do not execute logic in nfs hooks if netroot is not nfs
If the nfs module is present in the initrd, it always perform some operations
(mount/umount rpc_pipefs, exec/kill rpcbind, rpc.statd and rpc.idmapd), even in
initrds not specifically created for nfsroot (e.g., in non-hostonly mode).

This can create a conflict with some VMware tools that use the rpc_pipefs
filesystem and can run in the initrd.