feat: remove rev_lib_symlinks
rev_lib_symlinks is not used any more by dracut and it is not
a dependency for any of the out-of-tree dracut modules.

`rev_lib_symlinks` function was introduced by a10a141 to be used by
`inst_library`, however `inst_library` no longer uses it
(neither any other function).

It is hard to maintain code that is not documented,
not used and has no test coverage.