chore: move net udev rules to net-lib dracut modules
The net-lib dracut module is designed to be always installed when
networking is enabled in the generated initramfs.

Move networking related udev rules to net-lib dracut modules from the
udev-rules dracut module.

The motivation behind this change is to increase modularity,
stay consistent with other dracut modules and potentially
increase performance when networking is not required in the
initramfs.